refactor(repo): catalog admin CRUD 별도 module 로 분리
PR #44 의 후속. Phase 4.6 의 admin CRUD 5 메서드를 McpCatalogRepository
에서 추출 → McpCatalogAdminRepository 신규 module.

# 동기
- mcp-catalog-repository.ts 의 max-lines warning 완전 해소
  (522 → 497 → 384 lines, max 400 미만)
- 책임 분리 명확화:
  - McpCatalogRepository: 사용자용 read-only + from-catalog 등록 + instance
    lifecycle (격리)
  - McpCatalogAdminRepository: admin CRUD (disabled 포함)
  - McpAdminMonitoringRepository: admin observability (전역 aggregate)

# 변경
- mcp-catalog-admin-repository.ts 신규 (137 lines)
  - listAllForAdmin / getCatalogTemplateForAdmin / insertCatalogTemplate
    / updateCatalogTemplate / deleteCatalogTemplate
  - InsertCatalogTemplateInput / UpdateCatalogTemplatePatch type 명시
- mcp-catalog-repository.ts: admin block 제거 + pointer comment
- routes/mcp-catalog-admin.routes.ts: import + new ... 모두 교체

# 검증
- tsc 0 / lint 0 errors / 0 warnings (catalog repo)
- 전체 회귀: 97 suites / 1604 tests / 0 failed (동일)

# 후속
- skill-repository.ts 의 504 lines warning 은 별개 작업 후보
