+++ backend/api/src/routes/mcp-catalog-admin.routes.ts
@@ -18,7 +18,7 @@ import {
     createCatalogTemplateSchema,
     updateCatalogTemplateSchema,
 } from '../schemas/mcp-catalog-admin.schema';
-import { McpCatalogRepository } from '../data/repositories/mcp-catalog-repository';
+import { McpCatalogAdminRepository } from '../data/repositories/mcp-catalog-admin-repository';
 import { getUnifiedDatabase } from '../data/models/unified-database';
 import { createLogger } from '../utils/logger';
 
@@ -31,14 +31,14 @@ mcpCatalogAdminRouter.use(requireAuth, requireAdmin);
 
 // GET /catalog — disabled 포함
 mcpCatalogAdminRouter.get('/catalog', asyncHandler(async (_req: Request, res: Response) => {
-    const repo = new McpCatalogRepository(getUnifiedDatabase().getPool());
+    const repo = new McpCatalogAdminRepository(getUnifiedDatabase().getPool());
     const templates = await repo.listAllForAdmin();
     res.json(success({ templates, total: templates.length }));
 }));
 
 // GET /catalog/:id — disabled 포함 단건
 mcpCatalogAdminRouter.get('/catalog/:id', asyncHandler(async (req: Request, res: Response) => {
-    const repo = new McpCatalogRepository(getUnifiedDatabase().getPool());
+    const repo = new McpCatalogAdminRepository(getUnifiedDatabase().getPool());
     const tpl = await repo.getCatalogTemplateForAdmin(req.params.id);
     if (!tpl) {
         res.status(404).json(notFound('catalog template'));
@@ -54,7 +54,7 @@ mcpCatalogAdminRouter.post('/catalog', asyncHandler(async (req: Request, res: Re
         res.status(400).json({ success: false, error: 'VALIDATION_FAILED', details: parsed.error.issues });
         return;
     }
-    const repo = new McpCatalogRepository(getUnifiedDatabase().getPool());
+    const repo = new McpCatalogAdminRepository(getUnifiedDatabase().getPool());
     try {
         const created = await repo.insertCatalogTemplate(parsed.data);
         logger.info(`catalog template created: ${created.id} (by admin)`);
@@ -76,7 +76,7 @@ mcpCatalogAdminRouter.put('/catalog/:id', asyncHandler(async (req: Request, res:
         res.status(400).json({ success: false, error: 'VALIDATION_FAILED', details: parsed.error.issues });
         return;
     }
-    const repo = new McpCatalogRepository(getUnifiedDatabase().getPool());
+    const repo = new McpCatalogAdminRepository(getUnifiedDatabase().getPool());
     const updated = await repo.updateCatalogTemplate(req.params.id, parsed.data);
     if (!updated) {
         res.status(404).json(notFound('catalog template'));
@@ -88,7 +88,7 @@ mcpCatalogAdminRouter.put('/catalog/:id', asyncHandler(async (req: Request, res:
 
 // DELETE /catalog/:id — 영구 삭제
 mcpCatalogAdminRouter.delete('/catalog/:id', asyncHandler(async (req: Request, res: Response) => {
-    const repo = new McpCatalogRepository(getUnifiedDatabase().getPool());
+    const repo = new McpCatalogAdminRepository(getUnifiedDatabase().getPool());
     const removed = await repo.deleteCatalogTemplate(req.params.id);
     if (!removed) {
         res.status(404).json(notFound('catalog template'));
